pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "numpy>=1.21.0", "Cython>=0.29.0"]
build-backend = "setuptools.build_meta"

[project]
name = "patchmatch-cython"
version = "0.1.0"
description = "High-performance PatchMatch implementation for image inpainting using Cython"
readme = "README.md"
authors = [
    {name = "Teriks", email = "Teriks999@gmail.com"},
]
license = "BSD-3-Clause"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Cython",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["image-processing", "inpainting", "patchmatch", "cython", "computer-vision"]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.21.0",
]

[project.optional-dependencies]
visualization = [
    "matplotlib>=3.5.0",
    "Pillow>=8.0.0",
]
dev = [
    "pytest>=6.0.0",
    "build>=0.8.0",
    "setuptools>=61.0",
    "wheel",
    "Cython>=0.29.0",
]

[project.urls]
Homepage = "https://github.com/Teriks/patchmatch-cython"
Repository = "https://github.com/Teriks/patchmatch-cython"
Documentation = "https://patchmatch-cython.readthedocs.io"
Issues = "https://github.com/Teriks/patchmatch-cython/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["patchmatch_cython*"]

[tool.setuptools.package-data]
patchmatch_cython = ["*.pyx", "*.pxd", "*.pyd", "*.so", "*.cpp"]

[tool.cibuildwheel]
# Build for Python 3.10-3.13 only (skip older versions)
build = "cp310-* cp311-* cp312-* cp313-*"

# Configure for different platforms
[tool.cibuildwheel.linux]
archs = ["x86_64", "aarch64"]

[tool.cibuildwheel.windows]
archs = ["AMD64"]

[tool.cibuildwheel.macos]
archs = ["x86_64", "arm64"]

# Test command to verify the build
test-command = "python -c 'import patchmatch_cython; print(\"Import successful\")'" 